% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosmk.R
\name{autosmk}
\alias{autosmk}
\title{Generate template for Snakemake & R workflow}
\usage{
autosmk(name, inputs = NULL, outputs = NULL, directory = ".",
  open = TRUE)
}
\arguments{
\item{name}{Name of the Snakemake rule and R function to be created}

\item{inputs}{A named list of inputs; see details below}

\item{outputs}{A named list of outputs; see details below}

\item{directory}{The directory to write the output files to}

\item{open}{Logical - should the generated files be opened for editing? Default \code{TRUE}}
}
\description{
Generate template for Snakemake & R workflow
}
\examples{
inputs <- c('cool_csv' = 'data/input.csv')
outputs <- c('amazing_figure' = 'figs/figure.png')
autosmk('example_rule', inputs, outputs, tempdir(), open = FALSE)

}
