% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autosmk.R
\name{sanitize_inputs}
\alias{sanitize_inputs}
\title{Sanitize the inputs to the main functions}
\usage{
sanitize_inputs(name, inputs = NULL, outputs = NULL, directory = ".")
}
\description{
Sanitize the inputs to the main functions
}
\keyword{internal}
